refactor: split economy commands into separate files
Split the monolithic economy.go file into dedicated files for each command:
- balance.go: economy balance command
- set.go: economy set command
- top.go: economy top command
- pay.go: economy pay command

This improves code organization, maintainability, and readability by separating concerns.
